Refactor: estado general con Redux, scrollHandler para vincular scroll de three con el resto de la app
