+++ package.json
@@ -12,6 +12,7 @@
     "@react-spring/three": "^9.7.3",
     "@react-three/drei": "^9.80.4",
     "@react-three/fiber": "^8.13.7",
+    "@reduxjs/toolkit": "^1.9.5",
     "@use-gesture/react": "^10.2.27",
     "autoprefixer": "10.4.15",
     "next": "13.4.19",
@@ -19,6 +20,7 @@
     "postcss": "8.4.28",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-redux": "^8.1.2",
     "tailwindcss": "3.3.3",
     "three": "^0.156.0"
   }
