+++ src/app/layout.js
@@ -1,5 +1,6 @@
-import Navbar from '@/components/layout/navbar'
-import './globals.css'
+import ReduxProvider from './store/reduxProvider';
+import Navbar from '@/components/layout/navbar';
+import './globals.css';
 
 export const metadata = {
   title: 'Perla Labs',
@@ -10,8 +11,11 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={'font-mts-regular'}>
-        <Navbar />
-        {children}
+        {/* Proveedor que administra el estado global a los componentes hijos */}
+        <ReduxProvider>
+          <Navbar />
+          {children}
+        </ReduxProvider>
       </body>
     </html>
   )
