+++ src/app/store/store.js
@@ -1,0 +1,10 @@
+import { configureStore } from '@reduxjs/toolkit';
+import scrollReducer from './features/scrollSlice';
+
+export default configureStore(
+  {
+    reducer: {
+      scroll: scrollReducer,
+    },
+  }
+);
